chore: breakdown giant 1000+ line notifications file into subtyped module files
