+++ src/shared/copy/notifications/index.ts
@@ -1,0 +1,19 @@
+export * from './common'
+
+export * from './accounts-users-orgs'
+export * from './alerts'
+export * from './annotations'
+export * from './auth-and-tokens'
+export * from './billing-checkout'
+export * from './buckets'
+export * from './dashboard'
+export * from './functions'
+export * from './ingest'
+export * from './limits'
+export * from './notebooks'
+export * from './onboarding'
+export * from './operator'
+export * from './secrets'
+export * from './tasks'
+export * from './template'
+export * from './variables-urlQueries'
